print number of points as decimal value instead of integer [#45]

+++ core/app_client.go
@@ -258,15 +258,15 @@ func (a appClient) sendFashionQuizNotifications(orgID string, appID string, user
 				}
 
 				if notifyFirstDailyQuiz {
-					points := 0
+					points := float64(0)
 					if score != nil {
-						points = int(score.Score - oldScore.Score)
+						points = score.Score - oldScore.Score
 					}
 					pointsString := "points"
 					if points == 1 {
 						pointsString = "point"
 					}
-					body := fmt.Sprintf("@%s just scored %d %s in today's Runway Genius. Can you outplay them?", username, points, pointsString)
+					body := fmt.Sprintf("@%s just scored %g %s in today's Runway Genius. Can you outplay them?", username, points, pointsString)
 
 					message := model.NotificationMessage{
 						OrgID: orgID,
